SERIAL_SENSOR/requirements.txt
pyserial
paho-mqtt



